Modularized audioplayer component
